backend/requirements.txt
# Django Core
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Caching and Background Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
celery[redis]==5.3.6

# API Documentation
drf-spectacular==0.27.1

# Utilities
python-dotenv==1.0.1
shortuuid==1.0.12

# Validation
jsonschema==4.21.1

# Development
django-debug-toolbar==4.3.0
ipython==8.21.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==23.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.40.5
